Create blogs
Made a tool that makes it easier for me to create blogs. All I have to do is create a folder with a blog.md file, a banner file, and all the images that I linked in the blog.md file. Then I can run ```python3 manage.py blog_creator``` which will ask me to create a title and link my blog folder. The blog is then automatically created, and uploaded into the db.